move single-use components to same folder
