+++ app/services.tsx
@@ -1,4 +1,3 @@
-import {Card} from '@/components/ui/card'
 import {
   faFileSignature,
   faHandHoldingHand,
